LoRaMac: query minimum tx datarate
ADR allows end devices to adapt to degraded RF conditions by dropping
the datarate until the server responds with a downlink packet in
response to the ADRACKReq bit being set. In the case when the network
connection has been lost, ADR will reduce the tx datarate to the minimum
possible value, then take no further action.

By adding the ability to query the minimum possible tx datarate,
applications can detect when this condition has occured, and take
appropriate action to rectify it. This functionality must be integrated
into the stack as the minimum datarate depends on the current value of
UplinkDwellTime, which is not available externally.

Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>