Don't memset the TXRX buffer when radio is set to receive mode (#1345)
* Don't memset the TXRX buffer when radio is set to receive mode

The 1262 doesn't do this memset, but the 1272/1276 do.
This means, in an interrupt, if the radio is put back in RX mode before the payload is extracted and handed to a callback, it is zero'd.
This causes bugs like https://github.com/zephyrproject-rtos/zephyr/pull/48765
They would prefer it be fixed here, as it is just a waste anyway - we don't do this in 1262, and don't need to do it here.

* Don't memset buffer here

Similar to SX1276 change, this memset is unnecessary and causes bugs since SX1262 doesn't do it.

[mani: cherry-picked from upstream]
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>